Add ir_webstats_rc to the project
This is an old module I found on github so I don't have to
actually write the login stuff and it conveniently wraps the endpoints
so I am not using raw requests.

In the future I will write some objects to wrap the responses from the
endpoints so it isnt a json dump every time.