Allow bot to self-restart
When the bot loses connection to discord it restarts itself, this means
it sometimes is not initialize, so lets let it re-initialize itself the
first time someone runs a command so that we don't have to manually do
everything. This will make one command every so often be pretty slow,
but not prohibitively.